+++ examples/versioned-api/build.ps1
@@ -2,3 +2,6 @@ npm install
 tsp compile . --config azure-devops.tspconfig.yaml
 tsp compile . --config github.tspconfig.yaml
 tsp compile . --config docfx.tspconfig.yaml
+tsp compile . --config sphinx.tspconfig.yaml
+tsp compile . --config naturaldocs.tspconfig.yaml
+tsp compile . --config mkdocs.tspconfig.yaml